requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
Django==2.0.6
djangorestframework==3.8.2
more-itertools==4.2.0
pluggy==0.6.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
py==1.5.4
pytest==3.6.2
pytest-django==3.3.2
pytz==2018.4
six==1.11.0
